+++ CHANGELOG.md
@@ -11,6 +11,12 @@ This release is built on top of [OpenTelemetry .NET Auto Instrumentation v1.16.0
 
 ### Added
 
+- Experimental support for effective configuration reporting using OpAmp client.
+- Distribution zip files contain `SPLUNK_VERSION` file. It contains the version
+  and source reference of Splunk Distribution of OpenTelemetry .NET Automatic
+  Instrumentation. `VERSION` file keeps the data for the upstream OpenTelemetry
+  .NET Auto Instrumentation distribution.
+
 ### Changed
 
 ### Deprecated